refactor external router
